[FEATURE] NEAT address monitor (Linux)
This commit lays the foundation for the multiple prefix connection strategy. It
contains a NEAT context-skeleton, as well as functionality for monitoring all
available addresses on a Linux-host. The resolver needs to know which addresses
are available so that we can send DNS queries on multiple interfaces.

A similar interface exists in Windows and probably on BSD, but I don't run those
OS' so I can't check. The addresses are stored in a list, which is OK for now. I
don't believe in premature optimisations.