[FEATURE] Add v1 of internal callback API
Events inside NEAT will happen asynchronously. One example is addresses, they
will be added, deleted and deprecated during the lifetime of an application
using NEAT. An example of a subscriber is the resolver, it must always know
which interfaces it can use to send DNS requests. And even though caching (of
server/interface at least) will be implemented, the cache will expire and we
must send requests over all interfaces again.

In order to handle these asynchronous events, an internal callback API is
needed. This commit adds a first version. We have defined multi events
(neat_core.h) and neat_internal_context contains an array that is NEAT_MAX_EVENT
large. Each array element is a list of callback structs, and when an event
happens, we iterate through the list of "subscribers" and call the callback. We
currently only have one type of callback-function, which accepts the context and
a void pointer as input. This is good enough for now.  Components are added
to/removed from the list of subscribers using neat_add/remove_event_cb
(neat_core.c).

One potential drawback of the current solution is that we make use of queue.h
for lists. This limits the number of lists one callback struct can be member of
to 0. It is currently not a problem, but might be so in the future.

This API, in whatever form we end up with, might be useful for applications
using NEAT and could be made public.