[FEATURE] Implement domain name resolving
This commit adds initial code for doing DNS lookups to NEAT. Nothing is passed
to library user yet, but domain names are looked up on all available interfaces
(filtered by address family preference). The code is quite straight forward and
not that interesting. The most time-consuming task was refactoring the code so
that the resolver wraps the context, instead of the other way around. This way,
it is easy to support multiple contexts.

Currently, v4 addresses are used to lookup A records only, and V6 AAAAA.

Next steps:

* Handle address deletes/deprecate correctly and make sure we free memory
correctly.
* Decide on what to pass to library user and how.
* Implement basic multi prefix check (domains should be resolved to same IP).

Future work:

* Replace ldns. It is not available for Windows and allocates/frees memory like
crazy.