Add support for resolving literals
This commit introduces adds support for resolving literals. If a literal is
found, we wait 100ms and then iterate through all source addresses. For each
address with a matching family, a result structure is created. The rest of the
code works as before. The code works fine, but there is lots of room for
improvement. For example:

* We don't do any connection check now. Maybe not so critical, happy eyeballs
...
* There is no filtering on interfaces, so for example bridge interfaces are
included.