Only return unique IPs
We currently send DNS requests to four public web servers. The results was never
checked against already completed resolves for one (interface, src_addr), we
just blindly added the results and returned them to the user.

This is a waste, since what we want is the unique addresses resolved on one
(interface, src_addr). This commit introduces a basic and quite inefficient
search function. For each DNS result, we compare it with every other DNS result
on that interface and only add it if it is unique. Due to the current data
structure used when resolving addresses, this requires a double for-loop. A
better structure should be implemented, but this works fine for now.