Force TCP for SOCK_STREAM
DNS results are inserted in the inverse order. So when we clone the result for a
SOCK_STREAM (to set protocol as SCTP), then this might get used when connect is
called. Since happy eyeballs is not in place yet, force TCP for now. The
wildcard-behavior breaks for example neat_basic sample.