Support multiple protocols in resolver
In NEAT, the chosen protocol will dictate the socktype. Also, we might be
interested in trying a number of different protocols (Happy Eyeballs). Which
protocols to try is specified by the user selecting different properties. Store
this list in the resolver and then clone the result for each desired protocol.