Use DNS servers read from resolv.conf
On start and when resolv.conf is updated, we update the list of available DNS
servers to include whatever is contained in resolv.conf. Any missin server is
removed dynamically.

So far, there is no async handling of DNS servers. I.e., if no nameserver is
available on startup, then resolving will just time out. This is ok until we get
a use case for waiting with starting resolving until resolv.conf is updated.