Free memory after timeout callback
After the timeout callback has been called (for example that we have got a DNS
reply), free the memory allocated to the request. All resolver_pairs are moved
to a common list in the resolver (to ease the development of for example a
cache), and actually freeing the request has to be delayed until the timer has
been closed.