Add lookup of flows based on src and dst addrs
Add src and dst sockaddr structs to the neat flow, so we can keep track
of addresses for each end of connections. These structs will probably
need to become neat_addr structs so multi path support can be
implemented correctly.

Adds a conveinance function for comparing sock addrs.

src and dst pointers need to be populated for this to be useful, this is
a first step.