Revert "Switch from TAILQ to LIST"
This reverts commit 9487ae127874400fb3f2981c576c4b8dc1f1101e.

Use TAILQ to get expected behavior when user inserts a request (should be placed
laste).

Also, fixes a bug in core. Earlier, we stopped the loop before we release the
resolver. This will in some cases cause an assertion in libuv, as we might try
to update a closed/stopped loop (for example the timer used in requests). Free
resolver before closing loop. One drawback is that there will be a memory leak
on exit, but I dont view that as critical.