Don't always preserve message boundaries on the receiver side.
On the receiver side make the message boundary preservation
controlable by a variable. This is always false, so we have
only streaming mode for now.
We need to add some jason magic to make it user controllable.
The handling of partial reads is inefficient, since it requires
copying of message parts. This can be optimized later.
The sender is is also missing right now.