Improve asynchronous address handling
When the DNS literal timeout was reduced to 1ms, a bug in how we
retrieve addresses asynchronous (i.e., the initial address dump on
Linux) was exposed. neat_resolver_timeout_shared() was sometimes called
before any source address was received, causing the resolve callback to
be called with NEAT_ERROR (which is not handled btw).

I have chosen to fix this issue by adding a flag which is set when the
inital address dump is done. Addresses are read synchronously on BSD, so
the flag is set immediatly. Until this flag is set, we will keep calling
neat_resolver_timeout_shared() with a given interval.

Another "bug" discovered when fixing this issue, is that resolving
domains/literals on machines that has no IP (for example right after
boot) will fail. However, I don't think handling this error is the
responsibility of the resolver. How to handle a resolve failure is up to
for example the application, shall we give up or keep trying?