resolver: Handle DNS errors
If we get an error in return from the DNS server (for example NXDOMAIN),
start the complete timer to prevent waiting too long. It could be that
this timer is too short, and should be tweaked a bit (or logic improved)
in case other servers will send a valid reply.